feat: register ui 1/2
